src/backend/requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.27
icalendar